// apps/admin/src/shared/api/generated/ja-korea/schemas/awardCreateRequest.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */

export interface AwardCreateRequest {
  awardedDate: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  awardName: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  awardingInstitution: string;
}