// apps/admin/src/shared/api/generated/ja-korea/schemas/certificationUpdateRequest.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */

export interface CertificationUpdateRequest {
  certifiedDate: string;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  content: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  certificationInstitution: string;
  version: number;
}