// apps/admin/src/shared/api/generated/ja-korea/schemas/certificationsParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */
import type { CertificationsSort } from './certificationsSort';

export type CertificationsParams = {
published?: boolean;
content?: string;
institution?: string;
certifiedFrom?: string;
certifiedTo?: string;
createdFrom?: string;
createdTo?: string;
sort?: CertificationsSort;
page?: number;
size?: number;
};